source/file-processing/requirements.txt
functions-framework==3.*
google-cloud-storage==3.1.0
google-cloud-pubsub==2.28.0
google-cloud-bigquery==3.30.0